Draft of do_epoch + config.ini
Created the config.ini file for the hyperparameters and also made a test script to pull out information from it. Have a draft of do_epoch function. I tested getting the top "networks" with the best fitness function results using the made up data and the logic seems good. All thats left is to code the mutation functions as well as final testing once I know how the SNN network is formatted.