FINALLY FIXED THE DAMN WEIGHT ISSUE!!!!!!!
Been running the code a couple of times and it seems like all of the weights are (to my knowledge) transferring properly. There is still one known error. Currently, if certain conditions aren't met in the layer (i.e. not enough neurons) I have it coded so that the function calls itself recursively until it picks a valid linear layer. There have been 2 instances where this leads to a timeout error (too many recursive calls) and the code crashing. I'm leaving the code as is for now since the issue does not happen at a frequent enough rate. However, I will change it if it ends up becoming an issue in the future.